Implement a type-safe default value definition for InputValue. #111

+++ src/validators/int_validators.rs
@@ -4,10 +4,10 @@ use crate::Value;
 /// Integer range validator
 pub struct IntRange {
     /// Minimum value, including this value.
-    pub min: i64,
+    pub min: i32,
 
     /// Maximum value, including this value.
-    pub max: i64,
+    pub max: i32,
 }
 
 impl InputValueValidator for IntRange {
@@ -30,7 +30,7 @@ impl InputValueValidator for IntRange {
 /// Integer less then validator
 pub struct IntLessThan {
     /// Less then this value.
-    pub value: i64,
+    pub value: i32,
 }
 
 impl InputValueValidator for IntLessThan {
@@ -53,7 +53,7 @@ impl InputValueValidator for IntLessThan {
 /// Integer greater then validator
 pub struct IntGreaterThan {
     /// Greater then this value.
-    pub value: i64,
+    pub value: i32,
 }
 
 impl InputValueValidator for IntGreaterThan {
@@ -93,7 +93,7 @@ impl InputValueValidator for IntNonZero {
 /// Integer equal validator
 pub struct IntEqual {
     /// equal this value.
-    pub value: i64,
+    pub value: i32,
 }
 
 impl InputValueValidator for IntEqual {
